requirements.txt
blinker==1.9.0
click==8.3.0
flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
numpy==2.3.4
pandas==2.3.3
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
werkzeug==3.1.3
